refactor(filters): extracts FilterChipGroupDef and scopeFilterGroup to filterTypes.ts
